% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dash.R
\name{dash_view}
\alias{dash_view}
\title{Open the dashboard in a browser}
\usage{
dash_view(repo = here("strava_repo"))
}
\arguments{
\item{repo}{Path to the Strava repository.}
}
\value{
Path to the index page, invisibly.
}
\description{
Open the dashboard in a browser
}
