% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_dynamic_dashboard.R
\name{qd_view_dashboard}
\alias{qd_view_dashboard}
\title{Open the Quarto dashboard in a browser}
\usage{
qd_view_dashboard(repo = here("strava_repo"))
}
\arguments{
\item{repo}{Path to the Strava repository.}
}
\value{
Path to the index, invisibly.
}
\description{
Open the Quarto dashboard in a browser
}
