% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_static_dashboard.R
\name{qs_view_dashboard}
\alias{qs_view_dashboard}
\title{Open the static Quarto dashboard in a browser}
\usage{
qs_view_dashboard(repo = here("strava_repo"))
}
\arguments{
\item{repo}{Path to the Strava repository.}
}
\value{
Path to the index page, invisibly.
}
\description{
Open the static Quarto dashboard in a browser
}
