% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto_dashboard.R
\name{qview_dashboard}
\alias{qview_dashboard}
\title{Open the Quarto dashboard in a browser}
\usage{
qview_dashboard(repo = here("strava_repo"))
}
\arguments{
\item{repo}{Path to the Strava repository.}
}
\value{
Path to the index, invisibly.
}
\description{
Open the Quarto dashboard in a browser
}
