% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard.R
\name{view_dashboard}
\alias{view_dashboard}
\title{Open the dashboard in a browser}
\usage{
view_dashboard(repo = here("strava_repo"))
}
\arguments{
\item{repo}{Path to the Strava repository.}
}
\value{
Path to the index, invisibly.
}
\description{
Open the dashboard in a browser
}
